+++ ErsatzTV.Application/Scheduling/Commands/CreateBlockGroupHandler.cs
@@ -13,7 +13,7 @@ public class CreateBlockGroupHandler(IDbContextFactory<TvContext> dbContextFacto
         CancellationToken cancellationToken)
     {
         await using TvContext dbContext = await dbContextFactory.CreateDbContextAsync(cancellationToken);
-        Validation<BaseError, BlockGroup> validation = await Validate(request);
+        Validation<BaseError, BlockGroup> validation = await Validate(dbContext, request);
         return await validation.Apply(profile => PersistBlockGroup(dbContext, profile));
     }
 
@@ -24,10 +24,23 @@ public class CreateBlockGroupHandler(IDbContextFactory<TvContext> dbContextFacto
         return Mapper.ProjectToViewModel(blockGroup);
     }
 
-    private static Task<Validation<BaseError, BlockGroup>> Validate(CreateBlockGroup request) =>
-        Task.FromResult(ValidateName(request).Map(name => new BlockGroup { Name = name, Blocks = [] }));
+    private static Task<Validation<BaseError, BlockGroup>> Validate(TvContext dbContext, CreateBlockGroup request) =>
+        ValidateName(dbContext, request).MapT(name => new BlockGroup { Name = name, Blocks = [] });
 
-    private static Validation<BaseError, string> ValidateName(CreateBlockGroup createBlockGroup) =>
-        createBlockGroup.NotEmpty(x => x.Name)
-            .Bind(_ => createBlockGroup.NotLongerThan(50)(x => x.Name));
+    private static async Task<Validation<BaseError, string>> ValidateName(
+        TvContext dbContext,
+        CreateBlockGroup createBlockGroup)
+    {
+        Validation<BaseError, string> result1 = createBlockGroup.NotEmpty(c => c.Name)
+            .Bind(_ => createBlockGroup.NotLongerThan(50)(c => c.Name));
+
+        int duplicateNameCount = await dbContext.BlockGroups
+            .CountAsync(ps => ps.Name == createBlockGroup.Name);
+
+        var result2 = Optional(duplicateNameCount)
+            .Where(count => count == 0)
+            .ToValidation<BaseError>("Block group name must be unique");
+
+        return (result1, result2).Apply((_, _) => createBlockGroup.Name);
+    }
 }