+++ ErsatzTV.Application/Scheduling/Commands/CreateDecoTemplateHandler.cs
@@ -13,7 +13,7 @@ public class CreateDecoTemplateHandler(IDbContextFactory<TvContext> dbContextFac
         CancellationToken cancellationToken)
     {
         await using TvContext dbContext = await dbContextFactory.CreateDbContextAsync(cancellationToken);
-        Validation<BaseError, DecoTemplate> validation = await Validate(request);
+        Validation<BaseError, DecoTemplate> validation = await Validate(dbContext, request);
         return await validation.Apply(profile => PersistDecoTemplate(dbContext, profile));
     }
 
@@ -24,15 +24,30 @@ public class CreateDecoTemplateHandler(IDbContextFactory<TvContext> dbContextFac
         return Mapper.ProjectToViewModel(decoTemplate);
     }
 
-    private static Task<Validation<BaseError, DecoTemplate>> Validate(CreateDecoTemplate request) =>
-        Task.FromResult(
-            ValidateName(request).Map(name => new DecoTemplate
-            {
-                DecoTemplateGroupId = request.DecoTemplateGroupId,
-                Name = name
-            }));
+    private static async Task<Validation<BaseError, DecoTemplate>> Validate(
+        TvContext dbContext,
+        CreateDecoTemplate request) =>
+        await ValidateDecoTemplateName(dbContext, request).MapT(name => new DecoTemplate
+        {
+            DecoTemplateGroupId = request.DecoTemplateGroupId,
+            Name = name
+        });
 
-    private static Validation<BaseError, string> ValidateName(CreateDecoTemplate createDecoTemplate) =>
-        createDecoTemplate.NotEmpty(x => x.Name)
-            .Bind(_ => createDecoTemplate.NotLongerThan(50)(x => x.Name));
+    private static async Task<Validation<BaseError, string>> ValidateDecoTemplateName(
+        TvContext dbContext,
+        CreateDecoTemplate request)
+    {
+        if (request.Name.Length > 50)
+        {
+            return BaseError.New($"Deco template name \"{request.Name}\" is invalid");
+        }
+
+        Option<DecoTemplate> maybeExisting = await dbContext.DecoTemplates
+            .FirstOrDefaultAsync(r => r.DecoTemplateGroupId == request.DecoTemplateGroupId && r.Name == request.Name)
+            .Map(Optional);
+
+        return maybeExisting.IsSome
+            ? BaseError.New($"A deco template named \"{request.Name}\" already exists in that deco template group")
+            : Success<BaseError, string>(request.Name);
+    }
 }